test: added subset test fixtures
