+++ packages/utils/tests/subset/__fixtures__/index.ts
@@ -1,7 +1,9 @@
+import { invalid } from './invalid';
 import { valid } from './valid';
 
 export * from './types';
 
 export const fixtures = {
   valid,
+  invalid,
 };
