+++ packages/utils/tests/subset/__fixtures__/types.ts
@@ -1,0 +1,9 @@
+export type TestCase = {
+  subset: any;
+  superSet: any;
+};
+
+export type SubSetFixtures = {
+  valid: TestCase[];
+  invalid: TestCase[];
+};
